requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
alembic
psycopg2-binary
scikit-learn
pandas
matplotlib
prophet
python-dotenv
sendgrid
requests